cli/internal/jsonstream: move to top-level "internal"
