+++ internal/jsonstream/display.go
