+++ internal/jsonstream/display_test.go
